Set ec->rooteditor members when loading a project
This is required for the "preview for projects" feature.
Note, that as of now this only works for *loading* (=opening) a project,
not when a project is created and/or files are added to a project.
I still need to figure out where to change what in order to achieve that.